+++ client/src/components/utils/throttle.ts
@@ -1,23 +1,6 @@
 // Copyright © 2023, SAS Institute Inc., Cary, NC, USA.  All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-export interface Deferred<T> {
-  promise: Promise<T>;
-  resolve: (value: T | PromiseLike<T>) => void;
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  reject: (reason?: any) => void;
-}
-
-export function deferred<T>() {
-  // eslint-disable-next-line @typescript-eslint/consistent-type-assertions
-  const deferred = {} as Deferred<T>;
-  deferred.promise = new Promise((resolve, reject) => {
-    deferred.resolve = resolve;
-    deferred.reject = reject;
-  });
-  return deferred;
-}
-
 /**
  * Run tasks with limited concurrency
  * @param tasks array of tasks to run
